requirements.txt
# Spot Optimizer Platform - Python Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# WEB FRAMEWORK
# =============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# =============================================================================
# DATABASE & ORM
# =============================================================================
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# =============================================================================
# DATA VALIDATION
# =============================================================================
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# =============================================================================
# ASYNC TASK QUEUE
# =============================================================================
celery==5.3.6
redis==5.0.1

# =============================================================================
# AWS SDK
# =============================================================================
boto3==1.34.34
botocore==1.34.34

# =============================================================================
# AUTHENTICATION & SECURITY
# =============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# =============================================================================
# ML/DATA SCIENCE
# =============================================================================
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# =============================================================================
# HTTP CLIENT
# =============================================================================
requests==2.31.0
httpx==0.26.0

# =============================================================================
# TESTING
# =============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# =============================================================================
# LINTING & FORMATTING
# =============================================================================
black==24.1.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv==1.0.0
pytz==2024.1
python-dateutil==2.8.2

# =============================================================================
# MONITORING & LOGGING
# =============================================================================
prometheus-client==0.19.0
structlog==24.1.0

# =============================================================================
# STRIPE (Optional - for billing)
# =============================================================================
stripe==8.0.0

# =============================================================================
# EMAIL
# =============================================================================
sendgrid==6.11.0

# =============================================================================
# KUBERNETES CLIENT (for Agent)
# =============================================================================
kubernetes==29.0.0
